Separate JWT functionality to 2 functions in a separate file
